start vitest convertion and convert connect spec to vitest
