+++ packages/network/package.json
@@ -10,8 +10,8 @@
     "clean-deps": "rimraf node_modules",
     "lint": "eslint --ext .js,.jsx,.ts,.tsx,.json, .",
     "test": "yarn test-unit",
-    "test-debug": "yarn test-unit --inspect-brk=5566",
-    "test-unit": "mocha --reporter mocha-multi-reporters --reporter-options configFile=../../mocha-reporter-config.json",
+    "test-debug": "npx vitest --inspect-brk --no-file-parallelism --test-timeout=0",
+    "test-unit": "vitest run",
     "test-watch": "yarn test-unit --watch",
     "tslint": "tslint --config ../ts/tslint.json --project ."
   },
@@ -32,13 +32,10 @@
     "@cypress/request-promise": "^5.0.0",
     "@packages/https-proxy": "0.0.0-development",
     "@packages/socket": "0.0.0-development",
-    "@packages/ts": "0.0.0-development",
     "@types/concat-stream": "1.6.0",
     "express": "4.21.0",
-    "mocha": "6.2.2",
-    "sinon": "7.3.1",
-    "sinon-chai": "3.7.0",
-    "typescript": "~5.4.5"
+    "typescript": "~5.4.5",
+    "vitest": "^3.2.4"
   },
   "files": [
     "lib"
